+++ naga/src/front/wgsl/lower/mod.rs
@@ -3305,14 +3305,23 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
         size_expr: Handle<ast::Expression<'source>>,
         ctx: &mut ExpressionContext<'source, '_, '_>,
         span: Span,
-    ) -> Result<crate::PendingArraySize, Error<'source>> {
+    ) -> Result<Handle<crate::Override>, Error<'source>> {
         let expr = self.expression(size_expr, ctx)?;
         match resolve_inner!(ctx, expr).scalar_kind().ok_or(0) {
             Ok(crate::ScalarKind::Sint) | Ok(crate::ScalarKind::Uint) => Ok({
                 if let crate::Expression::Override(handle) = ctx.module.global_expressions[expr] {
-                    crate::PendingArraySize::Override(handle)
+                    handle
                 } else {
-                    crate::PendingArraySize::Expression(expr)
+                    let ty = ctx.register_type(expr)?;
+                    ctx.module.overrides.append(
+                        crate::Override {
+                            name: None,
+                            id: None,
+                            ty,
+                            init: Some(expr),
+                        },
+                        span,
+                    )
                 }
             }),
             _ => Err(Error::ExpectedConstExprConcreteIntegerScalar(span)),