+++ naga/src/lib.rs
@@ -502,15 +502,6 @@ pub struct Scalar {
     pub width: Bytes,
 }
 
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "serialize", derive(Serialize))]
-#[cfg_attr(feature = "deserialize", derive(Deserialize))]
-#[cfg_attr(feature = "arbitrary", derive(Arbitrary))]
-pub enum PendingArraySize {
-    Expression(Handle<Expression>),
-    Override(Handle<Override>),
-}
-
 /// Size of an array.
 #[repr(u8)]
 #[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
@@ -521,7 +512,7 @@ pub enum ArraySize {
     /// The array size is constant.
     Constant(core::num::NonZeroU32),
     /// The array size is an override-expression.
-    Pending(PendingArraySize),
+    Pending(Handle<Override>),
     /// The array size can change at runtime.
     Dynamic,
 }